social_scrapers/requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# Social Media Scrapers - Dependencies
# ══════════════════════════════════════════════════════════════════════════════

# Core Dependencies (Shared by all scrapers)
supabase>=2.0.0
python-dotenv>=1.0.0

# Instagram Scraper Dependencies
apify-client>=1.0.0
pandas>=2.0.0

# TikTok Scraper Dependencies
playwright>=1.40.0
requests>=2.31.0
tenacity>=8.0.0

# ══════════════════════════════════════════════════════════════════════════════
# Installation Instructions
# ══════════════════════════════════════════════════════════════════════════════
#
# 1. Install all dependencies:
#    pip install -r requirements.txt
#
# 2. Install Playwright browsers (required for TikTok scrapers):
#    playwright install
#
# 3. Set up environment variables:
#    - Copy .env.example to .env
#    - Fill in your API keys and credentials
#
# ══════════════════════════════════════════════════════════════════════════════
